Count Substrings

+++ codechef/PCSUB.cpp
@@ -9,7 +9,7 @@ int main() {
     while (t--) {
         cin >> n;
         string  s; cin >> s;
-        int count = 0;
+        long long int count = 0;
         for (int i=0; i<n; i++) {
             if (s[i] == '1')
                 count++;
